// Frontend/src/app/shared/model/user1.model.ts
export interface User1Model {
  accountNonExpired: boolean;
  accountNonLocked: boolean;
  authorities: [{}];
  createdAt: string;
  credentialsNonExpired: boolean;
  email: string;
  enabled: boolean;
  firstName: string;
  id: number;
  lastName: string;
  password: string;
  phone: string;
  role: string;
  username: string;
}